backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.34.2
boto3>=1.35.76
requests-oauthlib>=2.0.0
cryptography>=44.0.0
python-dotenv>=1.0.1
pymongo==4.9.1
pydantic>=2.10.3
email-validator>=2.2.0
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
tzdata>=2024.4
motor==3.6.0
pytest>=8.3.4
black>=24.10.0
isort>=5.13.2
flake8>=7.1.1
mypy>=1.14.0
python-jose[cryptography]>=3.3.0
requests>=2.32.3
python-multipart>=0.0.18
typer>=0.15.1
